+++ Src/Library/Extensions/HttpContextExtensions.cs
@@ -15,7 +15,7 @@ public static class HttpContextExtensions
         /// <typeparam name="TService">the type of the service to resolve</typeparam>
         /// <param name="keyName">the key name to resolve a keyed service</param>
         public TService? TryResolve<TService>(string keyName) where TService : class
-            => ServiceResolver.Instance.TryResolve<TService>(keyName);
+            => ServiceResolverClient.Forward.TryResolve<TService>(keyName);
 
         /// <summary>
         /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
@@ -23,7 +23,7 @@ public static class HttpContextExtensions
         /// <param name="typeOfService">the type of the service to resolve</param>
         /// <param name="keyName">the key name to resolve a keyed service</param>
         public object? TryResolve(Type typeOfService, string keyName)
-            => ServiceResolver.Instance.TryResolve(typeOfService, keyName);
+            => ServiceResolverClient.Forward.TryResolve(typeOfService, keyName);
 
         /// <summary>
         /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
@@ -32,7 +32,7 @@ public static class HttpContextExtensions
         /// <param name="keyName">the key name to resolve a keyed service</param>
         /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
         public TService Resolve<TService>(string keyName) where TService : class
-            => ServiceResolver.Instance.Resolve<TService>(keyName);
+            => ServiceResolverClient.Forward.Resolve<TService>(keyName);
 
         /// <summary>
         /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
@@ -41,21 +41,21 @@ public static class HttpContextExtensions
         /// <param name="keyName">the key name to resolve a keyed service</param>
         /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
         public object Resolve(Type typeOfService, string keyName)
-            => ServiceResolver.Instance.Resolve(typeOfService, keyName);
+            => ServiceResolverClient.Forward.Resolve(typeOfService, keyName);
 
         /// <summary>
         /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
         /// </summary>
         /// <typeparam name="TService">the type of the service to resolve</typeparam>
         public TService? TryResolve<TService>() where TService : class
-            => ServiceResolver.Instance.TryResolve<TService>();
+            => ServiceResolverClient.Forward.TryResolve<TService>();
 
         /// <summary>
         /// try to resolve an instance for the given type from the dependency injection container. will return null if unresolvable.
         /// </summary>
         /// <param name="typeOfService">the type of the service to resolve</param>
         public object? TryResolve(Type typeOfService)
-            => ServiceResolver.Instance.TryResolve(typeOfService);
+            => ServiceResolverClient.Forward.TryResolve(typeOfService);
 
         /// <summary>
         /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
@@ -63,7 +63,7 @@ public static class HttpContextExtensions
         /// <typeparam name="TService">the type of the service to resolve</typeparam>
         /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
         public TService Resolve<TService>() where TService : class
-            => ServiceResolver.Instance.Resolve<TService>();
+            => ServiceResolverClient.Forward.Resolve<TService>();
 
         /// <summary>
         /// resolve an instance for the given type from the dependency injection container. will throw if unresolvable.
@@ -71,7 +71,7 @@ public static class HttpContextExtensions
         /// <param name="typeOfService">the type of the service to resolve</param>
         /// <exception cref="InvalidOperationException">Thrown if requested service cannot be resolved</exception>
         public object Resolve(Type typeOfService)
-            => ServiceResolver.Instance.Resolve(typeOfService);
+            => ServiceResolverClient.Forward.Resolve(typeOfService);
 
         /// <summary>
         /// marks the current response as started so that <see cref="ResponseStarted(HttpContext)" /> can return the correct result.